requirements.txt
groq
flask
beautifulsoup4
requests
opencv-python
pytesseract
pillow
python-dotenv
